refactor: rename directory mock-data to mocks
