+++ project/src/mocks/favorite-locations.const.ts
