+++ project/src/mocks/hotels.const.ts
